+++ torch_geometric/nn/models/__init__.py
@@ -23,6 +23,7 @@ from .lightgcn import LightGCN
 from .mask_label import MaskLabel
 from .rev_gnn import GroupAddRev
 from .gnnff import GNNFF
+from .esan import DSnetwork, DSSnetwork
 from .pmlp import PMLP
 
 __all__ = classes = [
@@ -63,5 +64,7 @@ __all__ = classes = [
     'MaskLabel',
     'GroupAddRev',
     'GNNFF',
+    'DSnetwork',
+    'DSSnetwork',
     'PMLP',
 ]